+++ R/formatters.R
@@ -152,8 +152,13 @@ colformat_double <- function(
 #' @export
 #' @title format numeric cells
 #' @description Format numeric cells in a flextable.
+#'
+#' The function is different from [colformat_double()] on numeric type
+#' columns. The function uses the [format()] function of R on numeric
+#' type columns. So this is normally what you see on the R console
+#' most of the time (but scientific mode is disabled, NA are replaced, etc.).
 #' @inheritParams colformat_char
-#' @param big.mark,digits,decimal.mark see [format()]
+#' @param big.mark,decimal.mark see [format()]
 #' @family cells formatters
 #' @examples
 #' dat <- mtcars
@@ -161,7 +166,7 @@ colformat_double <- function(
 #' ft <- flextable(head(dat))
 #' ft <- colformat_num(x = ft,
 #'   big.mark=" ", decimal.mark = ",",
-#'   digits = 2, na_str = "N/A")
+#'   na_str = "N/A")
 #' ft <- autofit(ft)
 #' ft
 #' @section Illustrations:
@@ -171,7 +176,6 @@ colformat_num <- function(
   x, i = NULL, j = NULL,
   big.mark = get_flextable_defaults()$big.mark,
   decimal.mark = get_flextable_defaults()$decimal.mark,
-  digits = get_flextable_defaults()$digits,
   na_str = get_flextable_defaults()$na_str,
   prefix = "", suffix = ""){
 
@@ -180,11 +184,11 @@ colformat_num <- function(
 
   quo_fun <- quo(format_fun.default(
     x, big.mark = big.mark, decimal.mark = decimal.mark,
-    digits = digits, na_str = na_str, prefix = prefix, suffix = suffix
+    na_str = na_str, prefix = prefix, suffix = suffix
   ))
   fun_ <- new_function(
     pairlist2(x = , big.mark = big.mark, decimal.mark = decimal.mark,
-              digits = digits, na_str = na_str, prefix = prefix, suffix = suffix),
+              na_str = na_str, prefix = prefix, suffix = suffix),
     get_expr(quo_fun))
 
   docall_display(col_keys, fun_, x, i = i)