+++ NEWS.md
@@ -9,7 +9,10 @@ use `.` and loop over columns
 ## Issues
 
 * fix line spacing with pdf output
-
+* Now `colformat_num` calls the `format` function on the numeric values 
+(integer and float) which are therefore displayed as in console R. This function
+is used during the creation of a flextable so that by default the content of the
+cells is the same as that displayed in console R.
 
 # flextable 0.6.2